ingestionService/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
idna==2.10
jieba3k==0.35.1
joblib==1.0.1
lxml==4.6.3
newspaper3k==0.2.8
nltk==3.5
pika==1.2.0
Pillow==8.2.0
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.59.0
urllib3==1.26.4
